+++ qm/matrix.h
@@ -5,7 +5,7 @@
 ║                                                                                ║
 ║ PURPOSE:                                                                       ║
 ║   Custom templated matrix and vector classes supporting both real              ║
-║   (long double) and complex (std::complex<long double>) arithmetic.            ║
+║   (double) and complex (std::complex<double>) arithmetic.            ║
 ║   Specifically optimized for quantum mechanical calculations.                  ║
 ║                                                                                ║
 ║ KEY CLASSES:                                                                   ║
@@ -82,10 +82,10 @@ inline auto scalar_abs2(const T& x) -> decltype(std::real(x * scalar_conj(x)))
 
 // ─────────────────────────────────────────────────────────────────────────────
 // Templated vector<T>
-//   T = long double               for real calculations
-//   T = std::complex<long double> for complex (W-operator blocks, P-wave)
+//   T = double               for real calculations
+//   T = std::complex<double> for complex (W-operator blocks, P-wave)
 // ─────────────────────────────────────────────────────────────────────────────
-template<typename T = long double>
+template<typename T = double>
 struct vector {
     using value_type = T;
     std::vector<T> data;
@@ -173,8 +173,8 @@ T dot_no_conj(const vector<T>& v, const vector<T>& u) {
 }
 
 // Approximate equality (real types only)
-inline bool approx(long double x, long double y,
-                   long double acc = 1e-6, long double eps = 1e-6) {
+inline bool approx(double x, double y,
+                   double acc = 1e-6, double eps = 1e-6) {
     if (std::fabs(x - y) < acc) return true;
     if (std::fabs(x - y) < eps * (std::fabs(x) + std::fabs(y))) return true;
     return false;
@@ -192,10 +192,10 @@ std::ostream& operator<<(std::ostream& os, const vector<T>& v) {
 
 // ─────────────────────────────────────────────────────────────────────────────
 // Templated matrix<T>  (column-major: cols[j][i] = M(i,j))
-//   T = long double               for real matrices
-//   T = std::complex<long double> for complex matrices
+//   T = double               for real matrices
+//   T = std::complex<double> for complex matrices
 // ─────────────────────────────────────────────────────────────────────────────
-template<typename T = long double>
+template<typename T = double>
 struct matrix {
     using value_type = T;
     std::vector<vector<T>> cols;
@@ -502,8 +502,8 @@ std::ostream& operator<<(std::ostream& os, const matrix<T>& M) {
 // ─────────────────────────────────────────────────────────────────────────────
 // Convenience type aliases
 // ─────────────────────────────────────────────────────────────────────────────
-using ld      = long double;
-using cld     = std::complex<long double>;
+using ld      = double;
+using cld     = std::complex<double>;
 
 using rvec    = vector<ld>;   // real vector
 using cvec    = vector<cld>;  // complex vector
@@ -556,7 +556,7 @@ inline cmat to_complex(const rmat& A) {
     cmat R(A.size1(), A.size2());
     for (size_t i = 0; i < A.size1(); i++)
         for (size_t j = 0; j < A.size2(); j++)
-            R(i,j) = cld{A(i,j), 0.0L};
+            R(i,j) = cld{A(i,j), 0.0};
     return R;
 }
